fix(package): move ObjectSeeker to avoid package collision
